Do not create a cycle when adding control edges in ScopedAllocatorOptimizer.
ScopedAllocatorOptimizer adds a few control edges from the parent of the input
to collective ops.  More context: cl/265532843.

Before this change, the optimizer would not check if the node from which it was
adding a control edge to the ScopedAllocator node was already in the set of
inputs to the collective op group (input set).  If that was the case, this
would result in a cycle, because the optimizer also adds control edges from
ScopedAllocator node to all nodes in the input set.

After this change, the optimizer avoids adding a control edge if the node
belongs to the transitive fanout of the input set.

PiperOrigin-RevId: 280249308
Change-Id: Iacfb2543b2ac4ec4f0f0966ea73c903ab507cdaa