[XLA] Use dataflow analysis instead of alias analysis for setting preset assmts.
We were previously using alias analysis to find all aliased HloValues for a
given HloBuffer for the HloPosition, which led to allocating the same buffer
multiple times. Instead, use dataflow analysis to get only the unique HloValue.
This is because there is one-to-one correspondance between HloPosition and
HloValue, but multiple different HloPositions can map to the same HloBuffer in
case we have aliasing due to while loops.

PiperOrigin-RevId: 280318809
Change-Id: I9c674911de3cb8582c078d7accedf627507e12ca