Avoid copying the resource name in ResourceMgr::Lookup() and friends.
Previously, a ResourceMgr::Container map was keyed by an `std::pair<uint64,
string>`, and constructing this pair to perform a lookup would necessary copy
the resource name string. This change switches the key to be an `std::pair<uint64,
StringPiece>` and switches the value to be a struct containing the
ResourceBase* and a pointer to an owned string. As a result, lookup operations no longer
need to copy the name when creating a key.

As a consequence of making the Container value a struct, this change also moves reference count management to the struct's constructor/destructor, which simplifies code that deletes containers.

PiperOrigin-RevId: 280513227
Change-Id: Ic625e3f46679c48d7a5d521fe0c887d17414367c