Add ability for Executor subclasses to override the synchronous Run() method.
This change enables executors that have a more efficient synchronous implementation to use that implementation instead of requiring a callback to be created, and performing atomic operations to notify completion. This change also updates DirectSession to invoke the synchronous Run() method when the session is configured to run the graph on the caller thread (e.g. by setting `ConfigProto.inter_op_parallelism_threads = -1`).

PiperOrigin-RevId: 281562211
Change-Id: I2e8a32a440a82a32516555be8b2f64b373c21289