Lazily construct state in CancellationManager.
This change optimizes the common case of creating and destroying a CancellationManager that never has any callbacks registered in it. The callback map and cancelled_notification are lazily constructed the first time a callback is registered. As a result, for unused CancellationManagers, the destructor does not need to execute ~Notification (which includes a lock acquisition) and ~FlatMap.

PiperOrigin-RevId: 282568404
Change-Id: I606e5d870d8ae785a7a963cc45938cbf3077b3ab